requirements.txt
streamlit
beautifulsoup4
requests
googlesearch-python
fpdf